@@ -19,45 +19,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.0.3",
-    "@angular/common": "~8.0.3",
-    "@angular/compiler": "~8.0.3",
-    "@angular/core": "~8.0.3",
-    "@angular/forms": "~8.0.3",
-    "@angular/platform-browser": "~8.0.3",
-    "@angular/platform-browser-dynamic": "~8.0.3",
-    "@angular/router": "~8.0.3",
-    "core-js": "^3.1.4",
-    "rxjs": "~6.5.2",
-    "zone.js": "~0.9.1"
+    "@angular/animations": "~9.0.7",
+    "@angular/common": "~9.0.7",
+    "@angular/compiler": "~9.0.7",
+    "@angular/core": "~9.0.7",
+    "@angular/forms": "~9.0.7",
+    "@angular/platform-browser": "~9.0.7",
+    "@angular/platform-browser-dynamic": "~9.0.7",
+    "@angular/router": "~9.0.7",
+    "core-js": "^3.6.4",
+    "rxjs": "~6.5.4",
+    "tslib": "^1.11.1",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.6",
-    "@angular-devkit/build-ng-packagr": "~0.800.6",
-    "@angular/cli": "~8.0.6",
-    "@angular/compiler-cli": "~8.0.3",
-    "@angular/language-service": "~8.0.3",
-    "@types/jasmine": "~3.3.13",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~12.0.10",
-    "codelyzer": "~5.1.0",
-    "husky": "~2.7.0",
-    "jasmine-core": "~3.4.0",
+    "@angular-devkit/build-angular": "~0.900.7",
+    "@angular-devkit/build-ng-packagr": "~0.900.7",
+    "@angular/cli": "~9.0.7",
+    "@angular/compiler-cli": "~9.0.7",
+    "@angular/language-service": "~9.0.7",
+    "@types/jasmine": "~3.5.9",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
+    "husky": "~4.2.3",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "ng-packagr": "^5.3.0",
-    "prettier": "~1.18.2",
-    "pretty-quick": "~1.11.1",
-    "protractor": "~5.4.0",
-    "ts-node": "~8.3.0",
-    "tsickle": "~0.37.0",
-    "tslib": "^1.10.0",
-    "tslint": "~5.18.0",
-    "tslint-consistent-codestyle": "1.15.1",
+    "karma": "~4.4.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.1",
+    "karma-jasmine": "~3.1.1",
+    "karma-jasmine-html-reporter": "^1.5.2",
+    "ng-packagr": "^9.0.0",
+    "prettier": "~1.19.1",
+    "pretty-quick": "~2.0.1",
+    "protractor": "~5.4.3",
+    "ts-node": "~8.7.0",
+    "tslint": "~6.1.0",
+    "tslint-consistent-codestyle": "1.16.0",
     "typescript": "~3.7.0"
   },
   "husky": {
