@@ -1,6 +1,6 @@
 {
   "name": "ngx-state-stack",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "license": "MIT",
   "description": "Angular state management in form of a stack, for example when using Breadcrumbs",
   "author": {
@@ -17,8 +17,8 @@
     "Breadcrumbs"
   ],
   "peerDependencies": {
-    "@angular/common": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "rxjs": "~6.5.2"
+    "@angular/common": "^9.0.0",
+    "@angular/core": "^9.0.0",
+    "rxjs": "~6.5.4"
   }
 }
